almost works
something causes some name to be null...???

+++ idobrin-aspnet-logic/DTOs/Item/ItemReturn.cs
@@ -2,4 +2,4 @@ using idobrin_aspnet_logic.DTOs.Products;
 
 namespace idobrin_aspnet_logic.DTOs.Item;
 
-public record ItemReturn(int Id, int Quantity, decimal Price, ProductReturn Product);
+public record ItemReturn(int Id, int Quantity, float Price, ProductReturn Product);
